+++ src/creasy/azure/__init__.py
@@ -1,0 +1,1 @@
+"""Azure DevOps Server provider. Isolated from creasy.gitlab."""